force browser reload when context code changes
