+++ src/App/functions/checkPoolForWETH.ts
@@ -1,6 +1,5 @@
 import { isWrappedNativeToken } from '../../ambient-utils/dataLayer';
-import { PoolIF } from '../../ambient-utils/types';
-import { PoolDataIF } from '../../contexts/ExploreContext';
+import { PoolDataIF, PoolIF } from '../../ambient-utils/types';
 
 // fn to determine if the pool in question has WETH
 export default function checkPoolForWETH(pool: PoolIF | PoolDataIF): boolean {
