+++ src/contexts/AuctionsContext.tsx
@@ -85,6 +85,11 @@ export interface AuctionIF {
 export const AuctionsContext = createContext({} as AuctionsContextIF);
 
 export const AuctionsContextProvider = (props: { children: ReactNode }) => {
+    if (import.meta.hot) {
+        import.meta.hot.accept(() => {
+            window.location.reload(); // Forces a full browser reload when context code changes
+        });
+    }
     const {
         activeNetwork: { chainId },
     } = useContext(AppStateContext);
