+++ src/contexts/Futa/FutaHomeContext.tsx
@@ -29,6 +29,11 @@ export const FutaHomeContextProvider = ({
 }: {
     children: ReactNode;
 }) => {
+    if (import.meta.hot) {
+        import.meta.hot.accept(() => {
+            window.location.reload(); // Forces a full browser reload when context code changes
+        });
+    }
     const [isActionButtonVisible, setIsActionButtonVisible] = useState(false);
     const [showTerminal, setShowTerminal] = useState(true);
     const [hasVideoPlayedOnce, setHasVideoPlayedOnce] = useState(false);
