+++ src/contexts/TokenContext.tsx
@@ -24,6 +24,11 @@ export interface TokenContextIF {
 export const TokenContext = createContext({} as TokenContextIF);
 
 export const TokenContextProvider = (props: { children: React.ReactNode }) => {
+    if (import.meta.hot) {
+        import.meta.hot.accept(() => {
+            window.location.reload(); // Forces a full browser reload when context code changes
+        });
+    }
     const {
         activeNetwork: { chainId },
     } = useContext(AppStateContext);
