+++ src/contexts/TradeTokenContext.tsx
@@ -51,6 +51,11 @@ export interface TradeTokenContextIF {
 export const TradeTokenContext = createContext({} as TradeTokenContextIF);
 
 export const TradeTokenContextProvider = (props: { children: ReactNode }) => {
+    if (import.meta.hot) {
+        import.meta.hot.accept(() => {
+            window.location.reload(); // Forces a full browser reload when context code changes
+        });
+    }
     const {
         isUserIdle,
         activeNetwork: { chainId },
